feat: add the Toast component
